requirements.txt
transformers
torch
pillow
